PICOZXLCD - unified display driver
